message_edit: Fix stale syntax upon cancelling upload in preview.
Previously, canceling an upload in preview mode &
returning to editing left the syntax unchanged
(e.g., still showing "Uploading"). This happened
because the `replaceFieldText` function does not
work when the text area's display is none.

To fix this, instead of hiding the textarea in preview
mode (which sets it's display to none), we apply a
negative z-index and 0 height to make it disappear
without explicitly setting its display to none.